% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecdata.R
\docType{data}
\name{ecdata}
\alias{ecdata}
\title{Expected concentration data}
\format{A \code{data.frame} with the expected concentration of 30 analytes
of one plate. It consists in 1 background sample, 3 positives controls
and 16 dilution points (first dilution is duplicated).

\itemize{
\item{\code{sample}}{: type of sample (background, control or standard).}
\item{\code{analyte}}{: analyte tested.}
\item{\code{ec}}{: expected concentration value.}
}}
\description{
Expected concentration data based on multiplex
immunoassays experiments.
}
\keyword{data}

